make malloc work for 32bit

+++ include/ma/internal.h
@@ -63,8 +63,8 @@ typedef int ma_mtx;
 #define MA_FLAGS_MASK (MA_PINUSE_FLAG | MA_LARGE_FLAG | MA_SMALL_FLAG)
 #define MA_SIZE_MASK (~MA_FLAGS_MASK)
 
-#define MA_HEADER_SIZE (sizeof(size_t))
-#define MA_FOOTER_SIZE (sizeof(size_t))
+#define MA_HEADER_SIZE (sizeof(uint64_t))
+#define MA_FOOTER_SIZE (sizeof(uint64_t))
 
 #define MA_MALLOC_ALIGN (_Alignof(max_align_t))
 #define MA_HALF_MALLOC_ALIGN (MA_MALLOC_ALIGN / 2)
@@ -150,7 +150,7 @@ enum ma_size_class {
 };
 
 struct ma_hdr {
-	size_t tag;
+	uint64_t tag;
 
 	struct ma_hdr *next;
 	struct ma_hdr *prev;
